renaming some file to facilitate PR about the score
